requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Molecular dynamics analysis
MDAnalysis>=2.4.0
mdtraj>=1.9.0
pytraj>=2.0.0

# Clustering and machine learning
scikit-learn>=1.0.0
scipy>=1.7.0

# Structure analysis
nglview>=3.0.0
py3Dmol>=1.8.0

# Parallel computing
joblib>=1.1.0
dask>=2021.10.0

# Data handling
h5py>=3.1.0
netCDF4>=1.5.0

# Plotting and visualization
plotly>=5.0.0
ipywidgets>=7.6.0

# Configuration and CLI
click>=8.0.0
pyyaml>=6.0
configparser>=5.0.0

# Development tools
tqdm>=4.62.0
loguru>=0.6.0

# Optional: GPU acceleration (uncomment if CUDA available)
# cupy-cuda11x>=10.0.0 